backend/requirements.txt
Flask==0.12.2
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.2
Flask-Login==0.4.0
Werkzeug==0.12.2
PyJWT==1.5.0
